hold clk32kg for BT/WLAN
BlueTooth needs the clk32kg clk to be on
to be able to work properly. This clock may go
off in OFF in off mode, and when the clk32kg clk
was introduce as a resource may go off when wlan
is switched off.

Turn on clk32kg clk if BT is on.

Change-Id: Iedd00fe680aa86d3e6b25a46e1aec66118491e67